backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL driver
redis>=4.5.2,<5.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
boto3==1.34.0
stripe==7.8.0
supabase==2.3.0
httpx>=0.24.0,<0.25.0
pydantic==2.5.0
pydantic-settings==2.1.0
slowapi==0.1.9
pillow==10.1.0
# python-magic==0.4.27  # Not available on Windows - using alternative
cryptography>=41.0.0
celery==5.3.4
celery[redis]==5.3.4
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0
# Windows-specific dependencies
email-validator==2.1.0
